+++ hutool-core/src/main/java/cn/hutool/core/util/ImageUtil.java
@@ -19,6 +19,7 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.URL;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Random;
 
 import javax.imageio.ImageIO;
@@ -27,6 +28,7 @@ import javax.imageio.stream.ImageInputStream;
 import javax.imageio.stream.ImageOutputStream;
 
 import cn.hutool.core.codec.Base64;
+import cn.hutool.core.convert.Convert;
 import cn.hutool.core.exceptions.UtilException;
 import cn.hutool.core.img.Img;
 import cn.hutool.core.io.FileUtil;
@@ -452,14 +454,14 @@ public class ImageUtil {
 		Assert.notNull(srcImageFile);
 		Assert.notNull(destImageFile);
 		Assert.isFalse(srcImageFile.equals(destImageFile), "Src file is equals to dest file!");
-		
+
 		final String srcExtName = FileUtil.extName(srcImageFile);
 		final String destExtName = FileUtil.extName(destImageFile);
-		if(StrUtil.equalsIgnoreCase(srcExtName, destExtName)) {
-			//扩展名相同直接复制文件
+		if (StrUtil.equalsIgnoreCase(srcExtName, destExtName)) {
+			// 扩展名相同直接复制文件
 			FileUtil.copy(srcImageFile, destImageFile, true);
 		}
-		
+
 		ImageOutputStream imageOutputStream = null;
 		try {
 			imageOutputStream = ImageIO.createImageOutputStream(destImageFile);
@@ -504,7 +506,7 @@ public class ImageUtil {
 			throw new IORuntimeException(e);
 		}
 	}
-	
+
 	/**
 	 * 图像类型转换：GIF=》JPG、GIF=》PNG、PNG=》JPG、PNG=》GIF(X)、BMP=》PNG<br>
 	 * 此方法并不关闭流
@@ -938,7 +940,7 @@ public class ImageUtil {
 	public static BufferedImage pressImage(Image srcImage, Image pressImg, int x, int y, float alpha) {
 		return Img.from(srcImage).pressImage(pressImg, x, y, alpha).getImg();
 	}
-	
+
 	/**
 	 * 给图片添加图片水印<br>
 	 * 此方法并不关闭流
@@ -1151,9 +1153,10 @@ public class ImageUtil {
 	public static BufferedImage toImage(byte[] imageBytes) throws IORuntimeException {
 		return read(new ByteArrayInputStream(imageBytes));
 	}
-	
+
 	/**
 	 * 将图片对象转换为Base64形式
+	 * 
 	 * @param image 图片对象
 	 * @param imageType 图片类型
 	 * @return Base64的字符串表现形式
@@ -1473,6 +1476,121 @@ public class ImageUtil {
 		} catch (IOException e) {
 			throw new IORuntimeException(e);
 		}
+	}
+
+	// -------------------------------------------------------------------------------------------------------------------- Color
+	/**
+	 * Color对象转16进制表示，例如#fcf6d6
+	 * 
+	 * @param color {@link Color}
+	 * @return 16进制的颜色值，例如#fcf6d6
+	 * @since 4.1.14
+	 */
+	public static String toHex(Color color) {
+		String R = Integer.toHexString(color.getRed());
+		R = R.length() < 2 ? ('0' + R) : R;
+		String G = Integer.toHexString(color.getGreen());
+		G = G.length() < 2 ? ('0' + G) : G;
+		String B = Integer.toHexString(color.getBlue());
+		B = B.length() < 2 ? ('0' + B) : B;
+		return '#' + R + G + B;
+	}
+
+	/**
+	 * 16进制的颜色值转换为Color对象，例如#fcf6d6
+	 * 
+	 * @param hex 16进制的颜色值，例如#fcf6d6
+	 * @return {@link Color}
+	 * @since 4.1.14
+	 */
+	public static Color hexToColor(String hex) {
+		return getColor(Integer.parseInt(StrUtil.removePrefix("#", hex), 16));
+	}
+
+	/**
+	 * 获取一个RGB值对应的颜色
+	 * 
+	 * @param rgb RGB值
+	 * @return {@link Color}
+	 * @since 4.1.14
+	 */
+	public static Color getColor(int rgb) {
+		return new Color(rgb);
+	}
+
+	/**
+	 * 将颜色值转换成具体的颜色类型 汇集了常用的颜色集，支持以下几种形式：
+	 * 
+	 * <pre>
+	 * 1. 颜色的英文名（大小写皆可）
+	 * 2. 16进制表示，例如：#fcf6d6或者$fcf6d6
+	 * 3. RGB形式，例如：13,148,252
+	 * </pre>
+	 * 
+	 * 方法来自：com.lnwazg.kit
+	 * 
+	 * @param colorName 颜色的英文名，16进制表示或RGB表示
+	 * @return {@link Color}
+	 * @sin 4.1.14
+	 */
+	public static Color getColor(String colorName) {
+		if (StrUtil.isBlank(colorName)) {
+			return null;
+		}
+		colorName = colorName.toUpperCase();
+
+		if ("BLACK".equals(colorName)) {
+			return Color.BLACK;
+		} else if ("WHITE".equals(colorName)) {
+			return Color.WHITE;
+		} else if ("LIGHTGRAY".equals(colorName) || "LIGHT_GRAY".equals(colorName)) {
+			return Color.LIGHT_GRAY;
+		} else if ("GRAY".equals(colorName)) {
+			return Color.GRAY;
+		} else if ("DARK_GRAY".equals(colorName) || "DARK_GRAY".equals(colorName)) {
+			return Color.DARK_GRAY;
+		} else if ("RED".equals(colorName)) {
+			return Color.RED;
+		} else if ("PINK".equals(colorName)) {
+			return Color.PINK;
+		} else if ("ORANGE".equals(colorName)) {
+			return Color.ORANGE;
+		} else if ("YELLOW".equals(colorName)) {
+			return Color.YELLOW;
+		} else if ("GREEN".equals(colorName)) {
+			return Color.GREEN;
+		} else if ("MAGENTA".equals(colorName)) {
+			return Color.MAGENTA;
+		} else if ("CYAN".equals(colorName)) {
+			return Color.CYAN;
+		} else if ("BLUE".equals(colorName)) {
+			return Color.BLUE;
+		} else if ("DARKGOLD".equals(colorName)) {
+			// 暗金色
+			return hexToColor("#9e7e67");
+		} else if ("LIGHTGOLD".equals(colorName)) {
+			// 亮金色
+			return hexToColor("#ac9c85");
+		} else if (StrUtil.startWith(colorName, '#')) {
+			return hexToColor(colorName);
+		} else if (StrUtil.startWith(colorName, '$')) {
+			// 由于#在URL传输中无法传输，因此用$代替#
+			return hexToColor("#" + colorName.substring(1));
+		} else {
+			// rgb值
+			final List<String> rgb = StrUtil.split(colorName, ',');
+			if (3 == rgb.size()) {
+				final Integer r = Convert.toInt(rgb.get(0));
+				final Integer g = Convert.toInt(rgb.get(1));
+				final Integer b = Convert.toInt(rgb.get(2));
+				if (false == ArrayUtil.hasNull(r, g, b)) {
+					return new Color(r, g, b);
+				}
+			} else {
+				return null;
+			}
+		}
+		return null;
 	}
 
 	/**