+++ hutool-core/src/test/java/cn/hutool/core/util/ImageUtilTest.java
@@ -57,13 +57,13 @@ public class ImageUtilTest {
 	@Ignore
 	public void pressTextTest() {
 		ImageUtil.pressText(//
-				FileUtil.file("e:/test2.png"), //
+				FileUtil.file("e:/face.jpg"), //
 				FileUtil.file("e:/test2_result.png"), //
-				"123456", Color.RED, //
-				new Font("黑体", Font.BOLD, 40), //
-				-45, //
-				180, //
-				1.0f);
+				"版权所有", Color.WHITE, //
+				new Font("黑体", Font.BOLD, 100), //
+				0, //
+				0, //
+				0.8f);
 	}
 
 	@Test